fix bugs

+++ VLUTUTORS/Areas/Admin/Controllers/ManageCourseController.cs
@@ -210,8 +210,11 @@ namespace VLUTUTORS.Areas.Admin.Controllers
         public IActionResult DeleteLesson([FromForm] int hdInput)
         {
             Khoadaotao khoadaotao = _context.Khoadaotaos.Where(p => p.IdBaiHoc == hdInput).FirstOrDefault();
-            string filePath = Path.Combine(this._environment.WebRootPath, "Files", khoadaotao.TenBaiHoc.Trim());
-            Directory.Delete(filePath, true);
+            if(khoadaotao.TaiLieu !=null)
+            {
+                string filePath = Path.Combine(this._environment.WebRootPath, "Files", khoadaotao.TenBaiHoc.Trim());
+                Directory.Delete(filePath, true);
+            }
             _context.Khoadaotaos.Remove(khoadaotao);
             _context.SaveChanges();
             TempData["Message"] = "Xóa thành công!";
